# src/api/models/CurrencyType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum CurrencyType {
    USD = 'USD',
    EUR = 'EUR',
    PLN = 'PLN',
}