# src/api/models/PaginatedResult_Transaction_.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Transaction } from './Transaction';
export type PaginatedResult_Transaction_ = {
    data: Array<Transaction>;
    total: number;
    limit: number;
    offset: number;
};