# src/api/models/Statistics.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CurrencyType } from './CurrencyType';
export type Statistics = {
    currency: CurrencyType;
    symbol: string;
    totalVolume: number;
    totalCost: number;
    currentValue: number;
    avgPrice: number;
    grossPL: number;
};