# src/api/models/TransactionSortableFields.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum TransactionSortableFields {
    ID = 'id',
    ACCOUNT_ID = 'accountId',
    CURRENCY = 'currency',
    SYMBOL = 'symbol',
    VOLUME = 'volume',
    OPEN_TIME = 'openTime',
    OPEN_PRICE = 'openPrice',
    MARKET_PRICE = 'marketPrice',
    PURCHASE_VALUE = 'purchaseValue',
    GROSS_PL = 'grossPL',
    CREATED_AT = 'createdAt',
}